requirements.txt
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
django-jsonfield
django-rest-swagger
django-phonenumber-field
phonenumbers
djoser